Create README.md
Journal of Soil, Plant and Environment is an international journal dedicated to the advancements in agriculture throughout the world. The goal of this journal is to provide a platform for scientists, students, academics, and engineers all over the world to promote, share, and discuss various new issues and developments in different areas of agricultural sciences. All manuscripts must be prepared in English and are subject to a rigorous and fair peer-review process. Accepted papers will appear online within 3 weeks followed by printed hard copies.